+++ tests/testthat/test-49-px.R
@@ -55,14 +55,6 @@ test_that("px runs without errors (data frame and parquet path)", {
   expect_runs_without_errors("BEXSTA_parquet")
 })
 
-test_that("px runs without errors (URL)", {
-  skip_if_offline(host = "bank.stat.gl")
-
-  px(input = "https://bank.stat.gl:443/sq/deed6eff-22dd-4b25-9b78-fca11875352b")
-
-  expect_true(TRUE)
-})
-
 test_that("px can run on an Excel workbook without a 'Data' sheet", {
   px(input = get_metadata_path("BEXSTA"))