+++ NEWS.md
@@ -6,6 +6,7 @@ across all other headings.
 - Add support for short TIMEVAL syntax. (#435)
 - Bugfix: Add MAP to variables2 instead of table2. ($409)
 - Add support for OFFICIAL_STATISTICS keyword. (#443)
+- Add support for direct link download URL as input to `px()`. (#445)
 
 # pxmake 0.18.0